feat: 위시리스트 상품 수량 변경 로직 추가

+++ src/main/java/gift/service/WishService.java
@@ -1,11 +1,11 @@
 package gift.service;
 
-import gift.dto.ItemResponse;
 import gift.dto.WishRequest;
 import gift.dto.WishResponse;
 import gift.entity.Item;
 import gift.entity.Member;
 import gift.entity.Wish;
+import gift.exception.AuthorizationException;
 import gift.exception.ItemNotFoundException;
 import gift.repository.ItemRepository;
 import gift.repository.WishRepository;
@@ -34,11 +34,7 @@ public class WishService {
         if (wishes.isEmpty()) {
             return List.of();
         }
-
-        List<Long> productIds = wishes.stream()
-            .map(Wish::getProductId)
-            .toList();
-
+        List<Long> productIds = wishes.stream().map(Wish::getProductId).toList();
         Map<Long, Item> productsMap = itemRepository.findAllByIdIn(productIds).stream()
             .collect(Collectors.toMap(Item::getId, Function.identity()));
 
@@ -56,15 +52,36 @@ public class WishService {
 
     @Transactional
     public WishResponse addWish(WishRequest request, Member member) {
-        Wish wish = new Wish(null, member.getId(), request.productId());
-        Wish savedWish = wishRepository.save(wish);
+        Wish newWish = new Wish(null, member.getId(), request.productId(), request.quantity());
+        Wish savedWish = wishRepository.save(newWish);
+
         Item item = itemRepository.findById(savedWish.getProductId())
             .orElseThrow(() -> new ItemNotFoundException("상품 정보를 찾을 수 없습니다."));
         return WishResponse.from(savedWish, item);
     }
 
     @Transactional
-    public void deleteWish(Long wishId) {
+    public void updateWishQuantity(Long wishId, int quantity, Member loginMember) {
+        Wish wish = wishRepository.findById(wishId)
+            .orElseThrow(() -> new ItemNotFoundException("수정할 위시 항목을 찾을 수 없습니다."));
+
+        if (!wish.getMemberId().equals(loginMember.getId())) {
+            throw new AuthorizationException("자신의 위시리스트만 수정할 수 있습니다.");
+        }
+
+        wish.setQuantity(quantity);
+        wishRepository.update(wish);
+    }
+
+    @Transactional
+    public void deleteWish(Long wishId, Member loginMember) {
+        Wish wish = wishRepository.findById(wishId)
+            .orElseThrow(() -> new ItemNotFoundException("삭제할 위시 항목을 찾을 수 없습니다."));
+
+        if (!wish.getMemberId().equals(loginMember.getId())) {
+            throw new AuthorizationException("자신의 위시리스트만 삭제할 수 있습니다.");
+        }
+
         wishRepository.deleteById(wishId);
     }
 }
